# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Calculator API",
    "description" : "An API to perform basic arithmetic operations.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.example.com",
    "description" : "Production server"
  } ],
  "paths" : {
    "/token" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Get JWT token",
        "operationId" : "getToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/token_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "JWT token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          }
        }
      }
    },
    "/calculate/add" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform addition operation on two numerical values",
        "operationId" : "addOperation",
        "parameters" : [ {
          "name" : "number1",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 10
          }
        }, {
          "name" : "number2",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 5
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Addition result",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/calculate/subtract" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform subtraction operation on two numerical values",
        "operationId" : "subtractOperation",
        "parameters" : [ {
          "name" : "number1",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 10
          }
        }, {
          "name" : "number2",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 5
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Subtraction result",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/calculate/multiply" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform multiplication operation on two numerical values",
        "operationId" : "multiplyOperation",
        "parameters" : [ {
          "name" : "number1",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 10
          }
        }, {
          "name" : "number2",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 5
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Multiplication result",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/calculate/divide" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform division operation on two numerical values",
        "operationId" : "divideOperation",
        "parameters" : [ {
          "name" : "number1",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 10
          }
        }, {
          "name" : "number2",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "example" : 5
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Division result",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CalculationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "CalculationResponse" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "number",
            "example" : 2
          }
        },
        "example" : {
          "result" : 2
        }
      },
      "token_body" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "example" : "user"
          },
          "password" : {
            "type" : "string",
            "example" : "pass"
          }
        }
      },
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string",
            "example" : "your.jwt.token.here"
          }
        },
        "example" : {
          "token" : "your.jwt.token.here"
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}